fix conflict

+++ LSAPI/Controllers/StoreController.cs
@@ -26,106 +26,147 @@ namespace LSAPI.Controllers
         {
             try
             {
+                if (pageNumber.HasValue && pageNumber < 0)
+                {
+                    return BadRequest("pageNumber phải là số dương");
+                }
+
+                if (pageSize.HasValue && pageSize < 0)
+                {
+                    return BadRequest("pageSize phải là số dương");
+                }
+                if (id < 0)
+                {
+                    return BadRequest("id phải là số dương");
+                }
+                if (status < 0)
+                {
+                    return BadRequest("status phải là số dương");
+                }
+                if (zoneId < 0)
+                {
+                    return BadRequest("zoneId phải là số dương");
+                }
+                if (walletId < 0)
+                {
+                    return BadRequest("walletId phải là số dương");
+                }
+                if (accountId < 0)
+                {
+                    return BadRequest("accountId phải là số dương");
+                }
                 var rs = await _storeService.GetStore(id, storeName, status, zoneId, walletId, accountId, pageNumber, pageSize);
                 return Ok(rs);
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 return BadRequest($"Xem Store thất bại: {ex.Message}");
             }
-           
+
         }
         [HttpPost()]
         public async Task<ActionResult<StoreResponse>> PostStore(StoreRequest request)
         {
             try
             {
-                var regex = new Regex("^[0-9]+$");
-                var regex2 = new Regex("^[a-zA-Z]+$");
-                if (!regex.IsMatch(request.StoreName))
+
+                if (request.Status <0)
                 {
-                    return BadRequest("Tên cửa hàng không hợp lệ");
+                    return BadRequest("Status phải là số dương");
                 }
-                if (!regex2.IsMatch(request.StorePhone))
+                if (request.ZoneId < 0)
                 {
-                    return BadRequest("Số điện thoại không hợp lệ");
+                    return BadRequest("zoneId phải là số dương");
                 }
-                if (request.TemplateId <= 0)
+                if (request.WalletId < 0)
                 {
-                    return BadRequest("TemplateId phải là số nguyên dương");
+                    return BadRequest("walletId phải là số dương");
                 }
-                if (request.AccountId <= 0)
+                if (request.AccountId < 0)
                 {
-                    return BadRequest("AccountId phải là số nguyên dương");
+                    return BadRequest("accountId phải là số dương");
                 }
-                if (request.ZoneId <= 0)
+                if (request.TemplateId < 0)
                 {
-                    return BadRequest("ZoneId phải là số nguyên dương");
+                    return BadRequest("TemplateId phải là số dương");
                 }
-                if (request.Status <= 0)
+                var regex = new Regex(@"^\w+@gmail\.com$");
+                if (!regex.IsMatch(request.StoreEmail))
                 {
-                    return BadRequest("Status không hợp lệ");
+                    return BadRequest("Email phải có địa chỉ tên miền @gmail.com");
                 }
-                if (request.WalletId <= 0)
-                {
-                    return BadRequest("WalletId phải là số nguyên dương");
-                }
+                //var regexTime = new Regex(@"^\d{1,2}:\d{2}:\d{2}$");
+                //if (!regexTime.IsMatch(request.OpenTime.ToString()) || !regexTime.IsMatch(request.CloseTime.ToString()))
+                //{
+                //    return BadRequest("Giờ mở cửa và đóng cửa phải có định dạng 'HH:mm:ss', ví dụ: '7:00:00'");
+                //}
+                //if (!TimeSpan.TryParse(request.OpenTime.Value.ToString(), out _))
+                //{
+                //    ModelState.AddModelError("$.openTime", "Giờ mở cửa không hợp lệ");
+                //}
+                //if (!TimeSpan.TryParse(request.CloseTime.Value.ToString(), out _))
+                //{
+                //    ModelState.AddModelError("$.closeTime", "Giờ đóng cửa không hợp lệ");
+                //}
+                //if (!ModelState.IsValid)
+                //{
+                //    return BadRequest(ModelState);
+                //}
                 var rs = await _storeService.CreateStore(request);
                 return Ok(rs);
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 return BadRequest($"Tạo Store thất bại: {ex.Message}");
             }
-            
+
         }
         [HttpPut()]
         public async Task<ActionResult<StoreResponse>> PutStore(int id, StoreRequest storeRequest)
         {
             try
             {
+                if (id == 0)
+                {
+                    return BadRequest("làm ơn hãy nhập id");
+                }
                 if (id <= 0)
                 {
-                    return BadRequest("Id phải là số nguyên dương");
+                    return BadRequest("id phải là số dương");
                 }
-                var regex = new Regex("^[0-9]+$");
-                var regex2 = new Regex("^[a-zA-Z]+$");
-                if (!regex.IsMatch(storeRequest.StoreName))
+                if (storeRequest.Status < 0)
                 {
-                    return BadRequest("Tên cửa hàng không hợp lệ");
+                    return BadRequest("Status phải là số dương");
                 }
-                if (!regex2.IsMatch(storeRequest.StorePhone))
+                if (storeRequest.ZoneId < 0)
                 {
-                    return BadRequest("Số điện thoại không hợp lệ");
+                    return BadRequest("zoneId phải là số dương");
                 }
-                if (storeRequest.TemplateId <= 0)
+                if (storeRequest.WalletId < 0)
                 {
-                    return BadRequest("TemplateId phải là số nguyên dương");
+                    return BadRequest("walletId phải là số dương");
                 }
-                if (storeRequest.AccountId <= 0)
+                if (storeRequest.AccountId < 0)
                 {
-                    return BadRequest("AccountId phải là số nguyên dương");
+                    return BadRequest("accountId phải là số dương");
                 }
-                if (storeRequest.ZoneId <= 0)
+                if (storeRequest.TemplateId < 0)
                 {
-                    return BadRequest("ZoneId phải là số nguyên dương");
+                    return BadRequest("TemplateId phải là số dương");
                 }
-                if (storeRequest.Status <= 0)
+                var regex = new Regex(@"^\w+@gmail\.com$");
+                if (!regex.IsMatch(storeRequest.StoreEmail))
                 {
-                    return BadRequest("Status không hợp lệ");
-                }
-                if (storeRequest.WalletId <= 0)
-                {
-                    return BadRequest("WalletId phải là số nguyên dương");
+                    return BadRequest("Email phải có địa chỉ tên miền @gmail.com");
                 }
                 var rs = await _storeService.UpdateStore(id, storeRequest);
                 return Ok(rs);
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 return BadRequest($"Update Store thất bại: {ex.Message}");
             }
-            
+
         }
 
 
@@ -134,18 +175,22 @@ namespace LSAPI.Controllers
         {
             try
             {
+                if (id == 0)
+                {
+                    return BadRequest("làm ơn hãy nhập id");
+                }
                 if (id <= 0)
                 {
-                    return BadRequest("Id phải là số nguyên dương");
+                    return BadRequest("id phải là số dương");
                 }
                 var rs = await _storeService.DeleteStore(id);
                 return Ok(rs);
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 return BadRequest($"Xóa Store thất bại: {ex.Message}");
             }
-            
+
         }
 
         [HttpGet("api/stores/count")]
